fix nits
